requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
starlette==0.35.1
httpx==0.27.0

# Async Task Queue
celery[redis]==5.3.4
# Note: redis client installed automatically by celery[redis]

# AI & ML
google-generativeai==0.8.3

# Video Processing
numpy<2.0.0
opencv-python==4.9.0.80
# Note: moviepy removed - using ffmpeg subprocess directly

# Vector Database (for RAG)
chromadb==0.4.22

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1

# Utilities
requests==2.31.0

# Google Drive Integration
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Model Context Protocol
mcp
google-api-python-client
google-auth-oauthlib
google-auth-httplib2==1.0.0

# Note: Acontext observability uses requests (already included above)
# The acontext SDK is optional - we use the REST API directly
